[#320] refactor: 리뷰 신고 Entity 위치 이동
